Large-ish commit.
+ Added (experimental) late-binding code and some caching of information
pertaining to types/methods/fields. Goal is to improve speed, but at this
stage I have no idea if it'll do anything. Immediate binding still applies to
'noclass' tagged objects. The 'binding' attribute specifies the type of
binding.
 
+ 'rename' attribute now applies to types. Still does not apply to fields for
a variety of personal reasons.
 
+ Added LREF_Init function. This must be called somewhere in the main source
file, either directly or indirectly. Cannot be called in the root of imported
source files since there is no guarantee that all types will be registered
with reflection when that code is executed.
 
+ 'constructor' attribute added to rename the constructor function used by
types.
 
* Some functionality regarding methods being able to be passed around as
  delegate objects may soon become deprecated. No guarantee either way on how
  long this side-effect will be around.